feat:selesai mengerjakan tutorial-3-gopud (#4)

+++ gopud/src/main/java/apap/tutorial/gopud/controller/RestoranController.java
@@ -1,65 +1,91 @@
 package apap.tutorial.gopud.controller;
 
+import java.util.Collections;
+import java.util.Comparator;
 import java.util.List;
+
+import apap.tutorial.gopud.model.MenuModel;
+import apap.tutorial.gopud.service.MenuService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 import apap.tutorial.gopud.model.RestoranModel;
 import apap.tutorial.gopud.service.RestoranService;
 
 @Controller
 public class RestoranController{
+    @Qualifier("restoranServiceImpl")
     @Autowired
     private RestoranService restoranService;
-    // URL mapping add
-    @RequestMapping("/restoran/add")
-    public String add(
-            //Request Parameter untuk dipass
-            @RequestParam(value = "idRestoran", required = true) String idRestoran,
-            @RequestParam(value = "nama", required = true) String nama,
-            @RequestParam(value = "alamat", required = true) String alamat,
-            @RequestParam(value = "nomorTelepon", required = true) Integer nomorTelepon,
-            Model model
-    ) {
-        // Membuat objek RestoranModel
-        RestoranModel restoran = new RestoranModel(idRestoran, nama, alamat, nomorTelepon);
 
-        //Memanggil service addRestoran
+    @Autowired
+    private MenuService menuService;
+
+    @RequestMapping("/")
+    public String home() {return "home";}
+
+    // URL mapping yang digunakan untuk mengakses halaman add restoran
+    @RequestMapping(value = "/restoran/add", method = RequestMethod.GET)
+    public String addRestoranFormPage(Model model){
+        RestoranModel newRestoran = new RestoranModel();
+        model.addAttribute("restoran", newRestoran);
+        return "form-add-restoran";
+    }
+
+    // URL mapping yang digunakan untuk submit form yang telah anda masukkan pada halaman add restoran
+    @RequestMapping(value = "/restoran/add", method = RequestMethod.POST)
+    public String addRestoranSubmit(@ModelAttribute RestoranModel restoran, Model model) {
         restoranService.addRestoran(restoran);
-
-        //Add variabel nama restoran ke "namaResto" untuk dirender
-        model.addAttribute("namaResto", nama);
-
-        //Return view template
+        model.addAttribute("namaResto", restoran.getNama());
         return "add-restoran";
     }
 
     // URL mapping view
-    @RequestMapping("/restoran/view")
+    @RequestMapping(path = "/restoran/view", method = RequestMethod.GET)
     public String view(
             // Request Parameter untuk dipass
-            @RequestParam(value = "idRestoran") String idRestoran, Model model
+            @RequestParam(value = "idRestoran") Long idRestoran, Model model
     ){
         // Mengambil objek RestoranModel yang dituju
-        RestoranModel restoran = restoranService.getRestoranByIdRestoran(idRestoran);
+        RestoranModel restoran = restoranService.getRestoranByIdRestoran(idRestoran).get();
 
         // Add model restoran ke "resto" untuk dirender
         model.addAttribute("resto", restoran);
+
+        List<MenuModel> menuList = menuService.findAllMenuByIdRestoran(restoran.getIdRestoran());
+        model.addAttribute("menuList", menuList);
 
         // Return view template
         return "view-restoran";
     }
 
+    // API yang digunakan untuk menuju halaman form change restoran
+    @RequestMapping(value = "restoran/change/{idRestoran}", method = RequestMethod.GET)
+    public String changeRestoranFormPage(@PathVariable Long idRestoran, Model model){
+        // Mengambil existing data restoran
+        RestoranModel existingRestoran = restoranService.getRestoranByIdRestoran(idRestoran).get();
+        model.addAttribute("restoran", existingRestoran);
+        return "form-change-restoran";
+    }
+
+    // API yang digunakan untuk submit form change restoran
+    @RequestMapping(value = "restoran/change/{idRestoran}", method = RequestMethod.POST)
+    public String changeRestoranFormSubmit(@PathVariable Long idRestoran, @ModelAttribute RestoranModel restoran, Model model){
+        RestoranModel newRestoranData = restoranService.changeRestoran(restoran);
+        model.addAttribute("restoran", newRestoranData);
+        return "change-restoran";
+    }
+
     // URL mapping viewAll
-    @RequestMapping("restoran/viewall")
+    @RequestMapping("restoran/view-all")
     public String viewall(Model model){
 
         //Mengambil semua objek RestoranModel yang ada
         List<RestoranModel> listRestoran = restoranService.getRestoranList();
+        Collections.sort(listRestoran);
 
         // Add model restoran ke "resto" untuk di render
         model.addAttribute("restoList", listRestoran);
@@ -70,12 +96,14 @@ public class RestoranController{
 
     //URL mapping view restoran menggunakan path variable
     @RequestMapping("restoran/view/id-restoran/{idRestoran}")
-    public String viewPath(@PathVariable("idRestoran") String idRestoran, Model model){
+    public String viewPath(@PathVariable("idRestoran") Long idRestoran, Model model){
         // Mengambil objek RestoranModel yang dituju
-        RestoranModel restoran = restoranService.getRestoranByIdRestoran(idRestoran);
+        RestoranModel restoran = restoranService.getRestoranByIdRestoran(idRestoran).get();
 
         // Add model restoran ke "resto" untuk dirender
         model.addAttribute("resto", restoran);
+        List<MenuModel> menuList = menuService.findAllMenuByIdRestoran(restoran.getIdRestoran());
+        model.addAttribute("menuList", menuList);
 
         // Return view template
         return "view-restoran";
@@ -83,9 +111,9 @@ public class RestoranController{
 
     //URL mapping ganti nomor telepon
     @RequestMapping("restoran/update/id-restoran/{idRestoran}/nomor-telepon/{nomorTelepon}")
-    public String updateTelepon(@PathVariable("nomorTelepon") Integer nomorTelepon, @PathVariable("idRestoran") String idRestoran, Model model){
+    public String updateTelepon(@PathVariable("nomorTelepon") Integer nomorTelepon, @PathVariable("idRestoran") Long idRestoran, Model model){
         // Mengambil objek RestoranModel yang dituju
-        RestoranModel restoran = restoranService.getRestoranByIdRestoran(idRestoran);
+        RestoranModel restoran = restoranService.getRestoranByIdRestoran(idRestoran).get();
         restoran.setNomorTelepon(nomorTelepon);
 
         // Add model restoran ke "resto" untuk dirender
@@ -97,20 +125,25 @@ public class RestoranController{
 
     // URL mapping delete resto
     @RequestMapping("restoran/delete/id/{idRestoran}")
-    public String updateTelepon(@PathVariable("idRestoran") String idRestoran, Model model){
+    public String deleteRestoran(@PathVariable("idRestoran") Long idRestoran, Model model){
         // Mengambil objek RestoranModel yang dituju
-        RestoranModel restoran = restoranService.getRestoranByIdRestoran(idRestoran);
+        RestoranModel restoran = restoranService.getRestoranByIdRestoran(idRestoran).get();
         if(restoran == null || idRestoran.equals("")){
             return "error";
         }
         else{
             // Add model restoran ke "resto" untuk dirender
-            model.addAttribute("resto", restoran);
+            if (restoran.getListMenu().isEmpty()) {
+                model.addAttribute("resto", restoran);
 
-            restoranService.deleteResto(restoran);
+                restoranService.deleteResto(restoran);
 
-            // Return view template
-            return "delete-resto";
+                // Return view template
+                return "delete-resto";
+            }
+            else{
+                return "delete-error";
+            }
         }
     }
 }
